chore: Move schema file
