+++ mobile/src/db/schema.ts
@@ -15,7 +15,7 @@ import type { PlayFromSource } from "~/stores/Playback/types";
 
 //? Need to export all schemas from a single file.
 export * from "~/modules/customization/font/core/schema";
-export * from "~/modules/customization/theme/schema";
+export * from "~/modules/customization/theme/core/schema";
 
 //#region Artist
 export const artists = sqliteTable("artists", {
